requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.12
django-celery-beat==2.2.1
django-enum-choices==2.1.4
django-redis==5.0.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
kombu==5.2.4
prompt-toolkit==3.0.36
psycopg2-binary==2.9.3
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.4.2
six==1.16.0
sqlparse==0.4.3
vine==5.0.0
wcwidth==0.2.6
